Giving more control to the user over fetching results.
The original code in Simpledb._select used a while loop to get the
next-token from SimpleDB and retrieve all the results possible for
that query(to a max of 250 I guess), leaving the user no choice to
either retrieve less, if required, or to retieve the results beyond 250,
'coz he does not have the next-token for retieving more results.
The next-token is received from simpledb but consumed internally
in the loop and never given to user.

In the new code the user will get a maximum of 100 results at a time,
along with the next_token, so that he can request the next 100, and
so on and so forth.

The return value is not just a list (type) of results but a dictionary
which looks like
{'items': [....100 results from simpledb...],'next_token': next_token}

e.g.:
the end user - instead of writing this
    result = results = domain.select('select * from SomeDomain')

writes something like
    result = domain.select('select * from SomeDomain')
    items, next_token = result['items'], result['next_token']
    while next_token:
        result = domain.select('select * from SomeDomain', next_token=next_token)
        items.append(result['items'])
        next_token = result['next_token']

A generator function can separately be written if the user finds it cumbersome
to repeat the result =... items=... next_token=...   process.

cheers
JV