Fixed "limit" expression.
Earlier version did not have "limit n" working in the queries,
it used to return all results.

Reason:
The simpledb.Query class has an attribute called 'limit'
and a method also called 'limit'.

The method name is changed to set_limit and other required changes
to use this are fixed.